Test cases updated with requirements
