+++ cfme/tests/configure/test_analysis_profiles.py
@@ -367,7 +367,7 @@ def test_analysis_profile_description_validation(analysis_profile_collection):
 def test_custom_role_modify_for_dynamic_product_feature(request, appliance, product_features):
     """
     Polarion:
-        assignee: ghubale
+        assignee: apagac
         initialEstimate: 1/12h
         caseimportance: high
         caseposneg: positive
