+++ cfme/tests/infrastructure/test_tenant_quota.py
@@ -184,11 +184,11 @@ def test_tenant_quota_enforce_via_lifecycle_infra(appliance, provider, set_roott
     assert provision_request.row.reason.text == "Quota Exceeded"
 
 
-@pytest.mark.rhv3
 # first arg of parametrize is the list of fixtures or parameters,
 # second arg is a list of lists, with each one a test is to be generated
 # sequence is important here
 # indirect is the list where we define which fixtures are to be passed values indirectly.
+@pytest.mark.rhv3
 @pytest.mark.parametrize('context', [ViaSSUI, ViaUI])
 @pytest.mark.parametrize(
     ['set_roottenant_quota', 'custom_prov_data', 'extra_msg', 'set_default'],
@@ -226,11 +226,11 @@ def test_tenant_quota_enforce_via_service_infra(request, appliance, context, set
     assert provision_request.row.reason.text == "Quota Exceeded"
 
 
-@pytest.mark.rhv2
 # first arg of parametrize is the list of fixtures or parameters,
 # second arg is a list of lists, with each one a test is to be generated
 # sequence is important here
 # indirect is the list where we define which fixtures are to be passed values indirectly.
+@pytest.mark.rhv2
 @pytest.mark.parametrize(
     ['set_roottenant_quota', 'custom_prov_data'],
     [
